# deps: updates wazero to 1.0.0