--- a/bindings/wasm/output.go
+++ b/bindings/wasm/output.go
@@ -73,7 +73,7 @@ func NewWasmOutput(logger logger.Logger) bindings.OutputBinding {
 		runtimeConfig: wazero.NewRuntimeConfig().
 			WithCloseOnContextDone(true),
 
-		// The below violate sand-boxing, but allow code to behave as expecteout.
+		// The below violate sand-boxing, but allow code to behave as expected.
 		moduleConfig: wazero.NewModuleConfig().
 			WithRandSource(rand.Reader).
 			WithSysWalltime().