This updates [wazero](https://wazero.io/) to [1.0.0](https://github.com/tetratelabs/wazero/releases/tag/v1.0.0) which begins our compatability promise.

On behalf of everyone in the community, I want to thank you for trying wazero before we became stable. We will not raise unsolicited pull requests anymore. That said, if any update gives you problems, please feel free to contact [us](https://wazero.io/community/) on slack or via a GitHub issue.

```bash
$ go test -run='^$' -bench '^Benchmark.*' ./bindings/wasm ./middleware/http/wasm -count=6 > old.txt
$ go test -run='^$' -bench '^Benchmark.*' ./bindings/wasm ./middleware/http/wasm -count=6 > new.txt
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/dapr/components-contrib/bindings/wasm
           │   old.txt   │              new.txt              │
           │   sec/op    │   sec/op     vs base              │
Example-12   12.83µ ± 3%   11.81µ ± 2%  -7.98% (p=0.002 n=6)

pkg: github.com/dapr/components-contrib/middleware/http/wasm
                          │   old.txt   │              new.txt              │
                          │   sec/op    │   sec/op     vs base              │
Native/rewrite/rewrite-12   578.5n ± 1%   573.2n ± 0%  -0.92% (p=0.004 n=6)
Tinygo/rewrite/rewrite-12   1.173µ ± 1%   1.164µ ± 1%  -0.77% (p=0.011 n=6)
Wat/rewrite/rewrite-12      1.016µ ± 2%   1.021µ ± 4%       ~ (p=0.660 n=6)
geomean                     883.4n        879.7n       -0.42%
```